Module Hvsock

module Af_common : sig ... end

module Af_hyperv : sig ... end

Low-level interface to the AF_HYPERV socket family available on Windows kernels.

module Af_vsock : sig ... end

Low-level interface to the AF_VSOCK socket family available on Linux kernels.

module Hyperkit : sig ... end

Low-level interface to the Hyperkit AF_VSOCK interface

module Socket : sig ... end

An interface for hypervisor sockets which hides as many of the differences between Linux, Windows and Mac as possible.